Addressed ktlint and build failures
